Moves CLI "init" failure test to integration tests
